pulled in all symbol lookup code into tmem_syms.c
